Mylib 0.10

+++ include/CacheManager.php
@@ -2,29 +2,30 @@
 
 class CacheManager {
 
-	const SFBZIP_FILE = 'sfbzip';
+	const SFBZIP_DIR = 'sfbzip';
 	private static $cacheDir = 'cache/';
 	private static $dlDir = 'cache/dl/';
 	/** Time to Live for download cache (in hours) */
 	private static $dlTtl = 1;
 
 
-	public static function cacheExists($id) {
-		return file_exists(self::$cacheDir . $id);
+	public static function cacheExists($action, $id) {
+		return file_exists( self::getPath($action, $id) );
 	}
 
-	public static function getCache($id, $compressed = true) {
-		$c = file_get_contents(self::$cacheDir . $id);
+	public static function getCache($action, $id, $compressed = false) {
+		$c = file_get_contents( self::getPath($action, $id) );
 		return $compressed ? gzinflate($c) : $c;
 	}
 
-	public static function setCache($id, $content, $compressed = true) {
-		return file_put_contents(self::$cacheDir . $id,
+	public static function setCache($action, $id, $content, $compressed = false) {
+		$file = self::getPath($action, $id);
+		return myfile_put_contents($file,
 			$compressed ? gzdeflate($content) : $content);
 	}
 
-	public static function clearCache($id) {
-		$file = self::$cacheDir . $id;
+	public static function clearCache($action, $id) {
+		$file = self::getPath($action, $id);
 		return file_exists($file) ? unlink($file) : true;
 	}
 
@@ -34,10 +35,7 @@ class CacheManager {
 	}
 
 	public static function setDlFile($fname, $fcontent) {
-		if ( !file_exists(self::$dlDir) ) {
-			mkdir(self::$dlDir);
-		}
-		return file_put_contents(self::$dlDir . $fname, $fcontent);
+		return myfile_put_contents(self::$dlDir . $fname, $fcontent);
 	}
 
 
@@ -46,18 +44,26 @@ class CacheManager {
 	}
 
 	/**
-	 * delete all files older than a hour
-	 */
+	Deletes all download files older than the time to live.
+	*/
 	public static function deleteOldDlFiles() {
 		$thresholdTime = time() - self::$dlTtl * 3600;
 		$dh = opendir(self::$dlDir);
 		if (!$dh) return;
 		while (($file = readdir($dh)) !== false) {
+			if ( $file{0} == '.' ) { continue; }
 			$fullname = self::$dlDir . $file;
 			if (filemtime($fullname) < $thresholdTime) {
 				unlink($fullname);
 			}
 		}
 		closedir($dh);
+	}
+
+	public static function getPath($action, $id) {
+		$subdir = $action . '/';
+		settype($id, 'string');
+		$subsubdir = $id{0} . '/';
+		return self::$cacheDir . $subdir . $subsubdir . $id;
 	}
 }
